// src/types/productStoreTypes.tsx
export interface IProductData {
  id: number;
  type: string;
  photo: string;
  photoMin: string[];
  name: string;
  title: string;
  text: string[];
  wbUrl: string;
  ozonUrl: string;
}

export interface IShopData {
  id: number;
  type: string;
  photo: string;
  photoMin: string[];
  name: string;
  title: string;
  text: string[];
  wbUrl: string;
  ozonUrl: string;
}